there are 24 IOMUX configs per IO tile.  There are 4 pins per IO tile. From AF output, it seems every 4th IOMUX (up until IOMUX16) relates to the same underlying pin.
When a pin is used, IOMUX[0-4] is set.  But we also need to
clear the corresponding bits in IOMUX[4-15].

Add temporary hack to accomplish this.

Without the hack, a simple verilog consisting of:

assign leds = {1'b0, clk, 1'b0, 1'b1, 1'b0, clk, 1'b0, clk};

Resulted in the pins assigned to clock being equiv to 1'b1.

With the hack, those pins are now showing a nice 8MHz signal,
and the one assigned to 1'b1 still shows 3v3