add hack for fasm output when dealing with a LUT
tiles.py kindly inverts and reverses the LUT to match the original value.
Due to our global "#nofmt!", this LUT value would be used literally –
causing the bitstream value to be inverted&reversed from what it should be.

wrap fmt/nofmt around LUT lines to deal with this.

With this, one can output fasm from a bitstream, fasm_pack & pack,
and get a bit identical bitstream.