Adjust logic when the tile has fewer bitstream columns than other tiles in the same floorplan column.
Grab the bits out of the entry data using the tile's bs-columns
rather than floorplan column width.

As this will need padding, ensure the padding is added to the
left of the bits, not the right.

With this an agm-unpack/agm-pack process on an AG10K bitstream
produces identical results (ignoring that agm-pack does not yet
lzw compress the bitstream)