Replace encoders+formatters with value_transformers that can be used for both (and for decoding).
This is primarily to simplify tile definitions, make it so
they can be represented in json (couldn't previously due to the
use of lambda values), and to ensure consistency (some formatters
weren't updated when their associated encoder was, etc).

Packing and unpacking works w/o regression.  There may be
some regressions in unpacking to fasm format.

Remove annotations - there is no replacement for this yet, but
there is an automated way to do that with the wire db.

Make slices optional, only use it on IO tiles where the BELs
haven't been created yet.